requirements.txt
paramiko
mysql.connector
psycopg2-binary
psycopg2
python-nmap
distro


